backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-groq
sentence-transformers
pypdf
faiss-cpu
langchain-huggingface
langchain-community
python-dotenv
langchain-classic
